% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP.R
\name{GP}
\alias{GP}
\title{Prediction for data split by rsample::vfold_cv "GP's core"}
\usage{
GP(fold_split, phenotype_name, algorithm)
}
\arguments{
\item{fold_split}{data split by rsample::vfold_cv}

\item{phenotype_name}{The name of the phenotype to be predicted}

\item{algorithm}{Algorithm used for prediction}
}
\description{
Prediction for data split by rsample::vfold_cv "GP's core"
}
