% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_performance.R
\name{GP_evaluation}
\alias{GP_evaluation}
\title{Calculation of R and RMSE}
\usage{
GP_evaluation(prediction_df)
}
\arguments{
\item{prediction_df}{dataframe of prediction result}
}
\description{
Calculation of R and RMSE
}
