% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_performance.R
\name{GP_performance}
\alias{GP_performance}
\title{Evaluating the result of TeaWAS::GP_cv}
\usage{
GP_performance(prediction_list)
}
\arguments{
\item{prediction_list}{List of return values for TeaWAS::GP_cv}
}
\description{
Evaluating the result of TeaWAS::GP_cv
}
