% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaWAS-package.R
\docType{package}
\name{TeaWAS-package}
\alias{TeaWAS}
\alias{TeaWAS-package}
\title{TeaWAS: GWAS and GP analysis of tea plant}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yusuke Fukuda \email{tonegawa07@gmail.com}

}
\keyword{internal}
