% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_cv.R
\name{GP_cv}
\alias{GP_cv}
\title{Genomic Prediction by k-fold cross validation}
\usage{
GP_cv(phenotype, genotype, nfold, nrepeat, algorithm, outputdir)
}
\arguments{
\item{phenotype}{dataframe of phenotype}

\item{genotype}{matrix of SNPs data}

\item{nfold}{Number of divisions for cross-validation}

\item{nrepeat}{Number of prediction iterations}

\item{algorithm}{Algorithm to be used}

\item{outputdir}{Directory to output results}
}
\description{
Genomic Prediction by k-fold cross validation
}
