% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_cv_repeat.R
\name{GP_cv_repeat}
\alias{GP_cv_repeat}
\title{Genomic Prediction per phenotype}
\usage{
GP_cv_repeat(
  phenotype_name,
  phenotype_data,
  genotype_matrix,
  nfold,
  nrepeat,
  algorithm,
  outputdir,
  output = TRUE
)
}
\arguments{
\item{phenotype_name}{name of each phenotype}

\item{phenotype_data}{dataframe of each phenotype}

\item{genotype_matrix}{Transposed genotype matrix with equal number of rows}

\item{nfold}{Number of divisions for cross-validation}

\item{nrepeat}{Number of prediction iterations}

\item{algorithm}{Algorithm to be used}

\item{outputdir}{Directory to output results}

\item{output}{TRUE or FALSE}
}
\description{
Genomic Prediction per phenotype
}
